+++ src/cli-ux/styled/header.ts
@@ -1,0 +1,6 @@
+import * as chalk from 'chalk'
+import {CliUx} from '../../index'
+
+export default function styledHeader(header: string) {
+  CliUx.ux.info(chalk.dim('=== ') + chalk.bold(header) + '\n')
+}