+++ src/interfaces/hooks.ts
@@ -26,16 +26,16 @@ export interface Hooks {
     return: void;
   };
   preupdate: {
-    options: {channel: string};
+    options: {channel: string, version: string};
     return: void;
   };
   update: {
-    options: {channel: string};
+    options: {channel: string, version: string};
     return: void;
   };
   'command_not_found': {
     options: {id: string; argv?: string[]};
-    return: void;
+    return: unknown;
   };
   'plugins:preinstall': {
     options: {